refactor: modularize settings and trace helpers
Refs #25
