+++ app/api/projects/[id]/route.ts
@@ -1,5 +1,5 @@
 import { getProject } from "@/src/db/queries";
-import type { Project } from "@/src/db/schema";
+import { serializeProject } from "@/src/lib/api-serializers";
 
 export const runtime = "nodejs";
 
@@ -12,23 +12,4 @@ export async function GET(_req: Request, { params }: Ctx) {
     return Response.json({ error: "project not found" }, { status: 404 });
   }
   return Response.json({ project: serializeProject(project) });
-}
-
-function serializeProject(project: Project) {
-  return {
-    id: project.id,
-    provider: project.provider,
-    githubRepoId: project.githubRepoId,
-    githubInstallationId: project.githubInstallationId,
-    owner: project.owner,
-    name: project.name,
-    fullName: project.fullName,
-    defaultBranch: project.defaultBranch,
-    cloneUrl: project.cloneUrl,
-    localPath: project.localPath,
-    status: project.status,
-    lastError: project.lastError,
-    createdAt: project.createdAt.getTime(),
-    updatedAt: project.updatedAt.getTime(),
-  };
 }
