+++ app/api/projects/route.ts
@@ -4,7 +4,7 @@ import { getLocalWorkspacesRoot } from "@/src/workspace/local";
 import { cloneGitHubRepository, CloneError } from "@/src/workspace/clone";
 import { listProjects } from "@/src/db/queries";
 import { listInstallationRepositories } from "@/src/github/app";
-import type { Project } from "@/src/db/schema";
+import { serializeProject } from "@/src/lib/api-serializers";
 
 export const runtime = "nodejs";
 
@@ -46,25 +46,6 @@ export async function POST(req: Request) {
     const status = err instanceof CloneError ? 400 : 500;
     return Response.json({ error: errorMessage(err) }, { status });
   }
-}
-
-function serializeProject(project: Project) {
-  return {
-    id: project.id,
-    provider: project.provider,
-    githubRepoId: project.githubRepoId,
-    githubInstallationId: project.githubInstallationId,
-    owner: project.owner,
-    name: project.name,
-    fullName: project.fullName,
-    defaultBranch: project.defaultBranch,
-    cloneUrl: project.cloneUrl,
-    localPath: project.localPath,
-    status: project.status,
-    lastError: project.lastError,
-    createdAt: project.createdAt.getTime(),
-    updatedAt: project.updatedAt.getTime(),
-  };
 }
 
 function errorMessage(err: unknown): string {
