+++ components/chat/tool-card.tsx
@@ -1,18 +1,16 @@
 "use client";
 
 import type { ChatAddToolApproveResponseFunction } from "ai";
+import type { ToolState } from "@/src/lib/trace";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import { DiffView } from "./diff-view";
-
-type ToolState =
-  | "input-streaming"
-  | "input-available"
-  | "approval-requested"
-  | "approval-responded"
-  | "output-available"
-  | "output-error"
-  | "output-denied";
+import {
+  isOkWriteFileOutput,
+  pickString,
+  safeStringify,
+  toolStateBadge,
+} from "./tool-display";
 
 interface ToolCardProps {
   name: string;
@@ -33,7 +31,7 @@ export function ToolCard({
   approvalId,
   onApproval,
 }: ToolCardProps) {
-  const { label, tone } = stateBadge(state);
+  const { label, tone } = toolStateBadge(state);
 
   return (
     <details
@@ -42,9 +40,9 @@ export function ToolCard({
         "bg-card/50 text-foreground ring-1 ring-border",
       )}
     >
-      <summary className="flex items-center justify-between gap-2 cursor-pointer select-none px-3 py-2">
-        <span className="flex items-center gap-2 min-w-0">
-          <span className="font-mono text-[11px] font-semibold truncate">
+      <summary className="flex cursor-pointer select-none items-center justify-between gap-2 px-3 py-2">
+        <span className="flex min-w-0 items-center gap-2">
+          <span className="truncate font-mono text-[11px] font-semibold">
             {name}
           </span>
           <span
@@ -58,7 +56,7 @@ export function ToolCard({
         </span>
       </summary>
 
-      <div className="px-3 pb-3 pt-1 space-y-2">
+      <div className="space-y-2 px-3 pb-3 pt-1">
         {name === "write_file" ? (
           <WriteFileBody input={input} output={output} state={state} />
         ) : (
@@ -130,33 +128,35 @@ function WriteFileBody({
 }) {
   const nextContent = pickString(input, "content");
   const relPath = pickString(input, "path");
-
   const showDiff =
-    state === "output-available" && isOkWriteFileOutput(output) && nextContent !== undefined;
+    state === "output-available" &&
+    isOkWriteFileOutput(output) &&
+    nextContent !== undefined;
 
   if (showDiff) {
-    const out = output as WriteFileOkOutput;
     return (
       <div className="space-y-2">
         <Section title="file">
-          <span className="font-mono text-[11px]">{out.path ?? relPath ?? "?"}</span>
+          <span className="font-mono text-[11px]">
+            {output.path ?? relPath ?? "?"}
+          </span>
         </Section>
-        {out.previousTruncated ? (
+        {output.previousTruncated ? (
           <div className="rounded bg-background/60 px-3 py-2 text-[11px] text-muted-foreground ring-1 ring-border">
-            Existing file was too large to diff inline ({out.bytesWritten} bytes written).
+            Existing file was too large to diff inline ({output.bytesWritten}{" "}
+            bytes written).
           </div>
         ) : (
           <DiffView
-            previous={out.previousContent ?? ""}
-            next={nextContent ?? ""}
-            newFile={out.existed === false}
+            previous={output.previousContent ?? ""}
+            next={nextContent}
+            newFile={output.existed === false}
           />
         )}
       </div>
     );
   }
 
-  // Fallback for pre-output states: just show the file path + content preview.
   return (
     <div className="space-y-2">
       {relPath && (
@@ -173,30 +173,6 @@ function WriteFileBody({
       )}
     </div>
   );
-}
-
-type WriteFileOkOutput = {
-  ok: true;
-  path?: string;
-  bytesWritten?: number;
-  existed?: boolean;
-  previousContent?: string;
-  previousTruncated?: boolean;
-};
-
-function isOkWriteFileOutput(value: unknown): value is WriteFileOkOutput {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    "ok" in value &&
-    (value as { ok: unknown }).ok === true
-  );
-}
-
-function pickString(value: unknown, key: string): string | undefined {
-  if (typeof value !== "object" || value === null) return undefined;
-  const v = (value as Record<string, unknown>)[key];
-  return typeof v === "string" ? v : undefined;
 }
 
 function Section({
@@ -221,52 +197,4 @@ function Section({
       <div className="font-mono text-[11px] leading-snug">{children}</div>
     </div>
   );
-}
-
-function stateBadge(state: ToolState): { label: string; tone: string } {
-  switch (state) {
-    case "input-streaming":
-      return {
-        label: "calling…",
-        tone: "bg-secondary text-muted-foreground",
-      };
-    case "input-available":
-      return {
-        label: "running…",
-        tone: "bg-secondary text-muted-foreground",
-      };
-    case "approval-requested":
-      return {
-        label: "needs approval",
-        tone: "bg-amber-500/15 text-amber-600 dark:text-amber-400",
-      };
-    case "approval-responded":
-      return {
-        label: "approved",
-        tone: "bg-secondary text-muted-foreground",
-      };
-    case "output-available":
-      return {
-        label: "done",
-        tone: "bg-emerald-500/15 text-emerald-600 dark:text-emerald-400",
-      };
-    case "output-error":
-      return {
-        label: "error",
-        tone: "bg-destructive/15 text-destructive",
-      };
-    case "output-denied":
-      return {
-        label: "denied",
-        tone: "bg-destructive/15 text-destructive",
-      };
-  }
-}
-
-function safeStringify(value: unknown): string {
-  try {
-    return JSON.stringify(value, null, 2);
-  } catch {
-    return String(value);
-  }
 }
